+++ packages/ui/src/Fetch/index.ts
@@ -1,9 +1,4 @@
-export {
-  Fetch,
-  type FetchConstructor,
-  type FetchProps,
-  type FetchUpdateWrapper,
-} from './Fetch.js';
+export { Fetch, type FetchConstructor, type FetchProps } from './Fetch.js';
 export {
   FetchShopifyPartial,
   type FetchShopifyPartialConstructor,